综合：后台feed列表改成微博列表

+++ wojilu.Controller/Microblogs/My/MicroblogController.cs
@@ -175,14 +175,14 @@ namespace wojilu.Web.Controller.Microblogs.My {
                 throw new NullReferenceException( lang( "exDataNotFound" ) );
             }
 
-            if (blog.User.Id != ctx.viewer.Id) {
-                throw new Exception( lang( "exNoPermission" ) );
+            if (blog.User.Id != ctx.viewer.Id || !ctx.viewer.IsAdministrator()) {
+                echoError( lang( "exNoPermission" ) );
+                return;
             }
 
             microblogService.Delete( blog );
 
             echoAjaxOk();
-
         }
 
     }
